requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.1
fake-useragent==1.5.0
aiohttp-socks==0.8.4
cryptography==42.0.5
psutil==5.9.8
schedule==1.2.0
pytz==2024.1
playwright
aiosqlite
gunicorn==21.2.0
